Added - async function AnalyzeResultAndTeamsOfLastFiveGames(team) prototype
Description:
Analyze the five last games of @param team, and for each game, analyze the opponents 5 games prior to that game,
to find out, for example, what their stats was, and what possible effect that might have had on the outcome of that game
that @param team was in. This way, we can see if @param team's last 5 games, were more or less flukes, or if
the opponents they met, were coming into that game strong, or in a slump so to speak. This will give us a more
in depth understanding of @param team's last 5 games.

For example; if @param team has won it's last 5 games, we want to see, if the opponents in those 5 games, were actually
in trends that were positive or negative. (This is data, that actually might be useful to plug into some AI/ML stuff
but that is still kind of out of my reach of knowledge at this point). Either way, we can then compare @param team,
(@param team, meaning, the team name passed as a string to AnalyzeResultAndTeamsOfLastFiveGames) and it's oppponents
PDO, Corsi, Shots on goal, Save percentages, that each opponent had at the time of their game. So we are analyzing
each opponent during the time and prior to that game, not the opponents current stats.